+++ src/Temporalio/Worker/WorkflowInstance.cs
@@ -2350,6 +2350,7 @@ namespace Temporalio.Worker
                             Arguments = { payloadConverter.ToPayloads(input.Args) },
                             RetryPolicy = input.Options.RetryPolicy?.ToProto(),
                             CancellationType = (Bridge.Api.WorkflowCommands.ActivityCancellationType)input.Options.CancellationType,
+                            IncludeArgumentsInMarker = input.Options.IncludeArgumentsInMarker,
                         };
                         if (input.Headers is IDictionary<string, Payload> headers)
                         {
@@ -2821,7 +2822,7 @@ namespace Temporalio.Worker
                     OperationName: input.Operation.Name,
                     Arg: input.Arg,
                     Options: new(),
-                    Headers: null));
+                    Headers: input.Headers));
 
             private Task SignalExternalWorkflowInternalAsync(
                 ISerializationContext.Workflow serializationContext,
@@ -3117,8 +3118,7 @@ namespace Temporalio.Worker
             public override Task<NexusWorkflowOperationHandle<TResult>> StartNexusOperationAsync<TResult>(
                 string operationName, object? arg, NexusWorkflowOperationOptions? options = null) =>
                 SystemNexusPayloadVisitor.IsSystemEndpoint(Options.Endpoint) ?
-                instance.outbound.Value.ScheduleSystemNexusOperationAsync<TResult>(
-                    new(Service, new(operationName, arg?.GetType() ?? typeof(object), typeof(TResult)), arg)) :
+                instance.StartSystemNexusOperationAsync<TResult>(Service, operationName, arg) :
                 instance.outbound.Value.ScheduleNexusOperationAsync<TResult>(new(
                     Service: Service,
                     ClientOptions: Options,
@@ -3146,8 +3146,7 @@ namespace Temporalio.Worker
             public override Task<NexusWorkflowOperationHandle<TResult>> StartNexusOperationAsync<TResult>(
                 string operationName, object? arg, NexusWorkflowOperationOptions? options = null) =>
                 SystemNexusPayloadVisitor.IsSystemEndpoint(Options.Endpoint) ?
-                instance.outbound.Value.ScheduleSystemNexusOperationAsync<TResult>(
-                    new(Service, ServiceDefinition.Operations[operationName], arg)) :
+                instance.StartSystemNexusOperationAsync<TResult>(Service, operationName, arg) :
                 instance.outbound.Value.ScheduleNexusOperationAsync<TResult>(new(
                     Service: Service,
                     ClientOptions: Options,